+++ packages/core/src/card/Card.tsx
@@ -9,13 +9,13 @@ export const Root = forwardRef<HTMLDivElement, JSX.IntrinsicElements["div"]>(
       {...props}
       className={classNames(
         "dark:bg-secondary-800 flex flex-col rounded-md border dark:border-none dark:text-white dark:shadow-md",
-        className
+        className,
       )}
       ref={forwardedRef}
     >
       {children}
     </div>
-  )
+  ),
 );
 
 // CardHeader Component
@@ -25,13 +25,13 @@ export const Header = forwardRef<HTMLDivElement, JSX.IntrinsicElements["div"]>(
       {...props}
       className={classNames(
         "px-3xl py-lg dark:border-secondary-800 border-b text-lg font-semibold",
-        className
+        className,
       )}
       ref={forwardedRef}
     >
       {children}
     </div>
-  )
+  ),
 );
 
 // CardBody Component
@@ -44,7 +44,7 @@ export const Body = forwardRef<HTMLDivElement, JSX.IntrinsicElements["div"]>(
     >
       {children}
     </div>
-  )
+  ),
 );
 
 // CardFooter Component
@@ -54,11 +54,11 @@ export const Footer = forwardRef<HTMLDivElement, JSX.IntrinsicElements["div"]>(
       {...props}
       className={classNames(
         "px-3xl py-lg dark:border-secondary-800 flex w-full space-x-2 border-t",
-        className
+        className,
       )}
       ref={forwardedRef}
     >
       {children}
     </div>
-  )
+  ),
 );