+++ packages/core/CHANGELOG.md
@@ -10,11 +10,13 @@
 
 ### Patch Changes
 
--[x] Improved Dialog, React Select, Input and Password Field. -[x] Added Textarea and Tag Field
+-[x] Improved Dialog, React Select, Input and Password Field. -[x] Added
+Textarea and Tag Field
 
 ## 0.1.1
 
 ### Patch Changes
 
 - Resolved the css bug
-- ecf0b55: Separated the RUI from the cloud repo, for ease of use and development
+- ecf0b55: Separated the RUI from the cloud repo, for ease of use and
+  development